microservices/code/requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.2.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.3
Flask==2.3.3
Flask-Cors==4.0.0
fsspec==2026.1.0
ftfy==6.3.1
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-aiplatform==1.134.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.8.0
google-cloud-videointelligence==2.18.0
google-crc32c==1.8.0
google-genai==1.60.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.1
language_tool_python==3.2.2
locate==1.1.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmengine==0.10.7
more-itertools==10.8.0
msgpack==1.1.2
mysql-connector-python==8.4.0
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.2.0
openai-whisper==20250625
packaging==26.0
pillow==11.3.0
platformdirs==4.5.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.7.0
scenedetect==0.6.7.1
sentry-sdk==1.45.0
setuptools==80.10.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
termcolor==3.3.0
tiktoken==0.12.0
timm==1.0.24
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.7
ultralytics-thop==2.0.18
Unidecode==1.3.8
urllib3==2.6.3
wcwidth==0.3.1
websockets==15.0.1
Werkzeug==2.3.7
wordfreq==3.1.1
yapf==0.43.0
yt-dlp==2025.12.8
